Update compile script
Updates the compile script to account for the new directory structure.
Now each new letter should start by copying the template directory and
naming it for the coverletter recipient. The names within the directory
should not be changed but the output pdf file will mathch the name of
the directory.